+++ tests/test_data/test_timeseries.py
@@ -1,6 +1,6 @@
 import unittest
 
-from tfts.datasets.timeseries import TimeSeriesSequence
+from tfts.data.timeseries import TimeSeriesSequence
 
 
 class TimeSeriesSequenceTest(unittest.TestCase):
