+++ tfts/data/__init__.py
@@ -1,4 +1,4 @@
-"""tfts datasets"""
+"""tfts data"""
 
 from .get_data import get_air_passengers, get_sine
 from .timeseries import TimeSeriesSequence
