+++ tfts/features/embedding_feature.py
@@ -1,0 +1,12 @@
+"""Time series embedding feature"""
+
+import logging
+from typing import List
+
+import tensorflow as tf
+
+logger = logging.getLogger(__name__)
+
+
+def add_embedding_feature(features, cardinalities: List[int], embedding_dims: List[int]):
+    return
