+++ tfts/models/__init__.py
@@ -1,4 +1,11 @@
 """tfts models"""
 
 from .auto_config import AutoConfig
-from .auto_model import AutoModel
+from .auto_model import (
+    AutoModel,
+    AutoModelForAnomaly,
+    AutoModelForClassification,
+    AutoModelForPrediction,
+    AutoModelForSegmentation,
+    AutoModelForUncertainty,
+)
